Remove some code that didn't prove fruitful for doing typed migrations
